清除time_t的warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@21084 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/bson/bson-inl.h
@@ -718,7 +718,8 @@ namespace bson {
             memset ( buffer, 0, 64 ) ;
             time_t timer = (time_t)( ( (long long)date() ) / 1000 ) ;
             //[ 1900-01-01, 9999-12-31 ]
-            if( timer >= -2208988800 && timer <= 253402300799 )
+            if( (long long)timer >= -2208988800LL &&
+                (long long)timer <= 253402300799LL )
             {
                local_time ( &timer, &psr ) ;
                sprintf ( buffer,
